🛠️ refactor: move PanelCommandsManager to the notifications directory
